requirements.txt
setuptools~=46.1.3
Pillow~=7.1.2
click~=7.1.1